cumulus: Remove `table X` option from static routes
Upstream has the ability to handle static routes that include
a table-id if we are working in a namespace based vrf.

Since the sub-command only works with namespaced based vrf's
and we don't support it, just remove the table X option
as that it will cause confusion that we don't want to have.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>