pimd: add allow-rp knob to ignore incorrect rp
When processing a (*,G) source list entry, the RFC dictates that the
source address provided must match the RP address. In some situations
it's desirable to forego this check. This patch adds a simple boolean
knob, configurable on a per-interface basis, to disable that check.

Alternatively, one can specify a prefix-list, which will act as a
whitelist for what RP addresses to allow.

Signed-off-by: Quentin Young <qlyoung@nvidia.com>