tools,watchfrr: coordinate timeout when restarting a daemon
When watchfrr restarts a hung or failed daemon, convey the
timeout that it's going to use to the shell script that's
actually doing the work. If the two timeouts are different,
watchfrr may terminate the shell script prematurely.

Also add a "kill" step to frrcommon's daemon_stop(): if
a daemon is hung or deadlocked, it may not be able to react
to SIGINT, and SIGKILL may be necessary.

Signed-off-by: Mark Stapp <mstapp@nvidia.com>