[CUMULUS] tools: Rearranging execution order for cmds upon frr-reload
Currently when we modify a router bgp nbr password and add the 'service
password-obfuscation' in frr.conf manually followed by a reload, the nbr
password command is executed before service password-obfuscation.

Because of this the passwords read from frr.conf by BGPd are assumed to
be original unencrypted passwords and an additional encrpytion is added.

Fix: Push the 'service password-obfuscation' cmd to top of line

Note: Issue is seen every time when we enable password-obfuscation from
NVUE.

Issue: 3500020

Ticket: 3500020

Signed-off-by: Chirag Shah <chirag@nvidia.com>
Signed-off-by: Rajasekar Raja <rajasekarr@nvidia.com>