[CUMULUS] zebra: Ensure WFI installs in a situation
When upgrading to latest dev the ability to notify interested
parties about a wait for install failure was lost due to the
upgrade re-arranging how code was handled in the dplane.
Effectively the dplane path was short-circuited for the case
where the new update is from the kernel to not send it back
to the kernel.

Ticket: #3486467
Signed-Off-By: Donald Sharp <sharpd@nvidia.com>