lib: When VRF is not initialized, show the target VRF as none
Issue:

When the switch is rebooted, and a script runs in parallel to get the
next hop groups upon switch startup, there could be a scenario where
the VRF is not completely initialized. In this case, we try to look up by
the ID and access the null pointer.

Fix:

When the VRF is VRF_UNKNOWN, print none instead of accessing the null VRF.
 Subsequently, execution of the next hop will show the correct VRF.

Ticket: #3497700

Signed-off-by: Rajesh Varatharaj <rvaratharaj@nvidia.com>