[CUMULUS]zebra: Cache and re-install rejected routes to kernel on NH install
Intermittently the NH installation in kernel fails and the routes are rejected
on quick link flaps.
On interface address add zebra does a re-installation of nexthop's which don't
have the 'INSTALL' flag set.
Walk the cache of rejected routes stored against the nhe and
install on successful install of the nexthop.

Ticket:#3428677
Testing Done: UT and test run on failing verification testbed (attached to MR)

Logs for rejected and re-installation:

Rejected route cache saved against NHE

2023-08-28T22:23:08.916988+00:00 mlx-3700-15 zebra[210270]: [TERH1-34712] Rejected route adding to NHE
2023-08-28T22:23:08.917084+00:00 mlx-3700-15 zebra[210270]: [VYKYC-709DP] vrf1(53:1002):133.1.1.0/24: Route install failed
2023-08-28T22:23:08.917165+00:00 mlx-3700-15 zebra[210270]: [TERH1-34712] Rejected route adding to NHE
2023-08-28T22:23:08.918151+00:00 mlx-3700-15 zebra[210270]: [HFAXM-2790E] vrf1(53):210.0.2.251/32: Evaluate RNH,

Re-install of the route on NHG install is sucessful

2023-08-28T22:23:08.947539+00:00 mlx-3700-15 zebra[210270]: [ZHFB4-0MJKD] zebra_interface_nhg_reinstall install nhe 4403[if 54] nh type 1 flags 0x5
2023-08-28T22:23:08.947632+00:00 mlx-3700-15 zebra[210270]: [NBGDS-5ZDQ9] MESSAGE: ZEBRA_INTERFACE_ADDRESS_ADD fe80::200:5eff:fe00:101/64 on vlan1001-v0 vrf vrf1(53)
2023-08-28T22:23:08.947722+00:00 mlx-3700-15 zebra[210270]: [ZHFB4-0MJKD] zebra_interface_nhg_reinstall install nhe 4403[if 54] nh type 1 flags 0x5
2023-08-28T22:23:08.947814+00:00 mlx-3700-15 zebra[210270]: [NCS3C-G7W12] Rejected route re-install on nhg install
2023-08-28T22:23:08.947918+00:00 mlx-3700-15 zebra[210270]: [NCS3C-G7W12] Rejected route re-install on nhg install
2023-08-28T22:23:08.948013+00:00 mlx-3700-15 zebra[210270]: [N7R9G-4ER1V] netlink_route_multipath_msg_encode: RTM_NEWROUTE 133.1.1.0/24 vrf 0(254) rtm_flags: 0x0
Signed-off-by: Ashwini Reddy <ashred@nvidia.com>

[CUMULUS] zebra: fix seg fault in nhg rej route installation from MR529

On rib route processing rn can be present without re's.
Hence check if re is NULL before accessing the re->nhe

prefixlen = 1025, ptr = 0}}}, table = 0x55de04ebbfc0,
parent = 0x55de0517b290, link = {0x0, 0x0}, lock = 4, nodehash = {hi = {next = 0x0,
      hashval = 2574579041}}, info = 0x0}

Was introduced with -
    https://gitlab-master.nvidia.com/nbu-sws/CL/FRR/frr/-/merge_requests/523

NOTE: this commit will not be required to be carried forward as
      if upstreamed we will include this fix in the original commit.

Ticket:#3593844

Testing:

UT: ran the bgp test that failed and also ran the reject route sequence
with theses changes
-
precommit that was failing which passed
https://jenkins-cl-qa-1.lab2300.labs.mlnx/blue/organizations/jenkins/user_jobs%2Fashred%2Fashred_job/detail/ashred_job/17/

Signed-off-by: Ashwini Reddy <ashred@nvidia.com>

[CUMULUS] zebra: fix seg fault in dbg msg introduced from MR523

When the route node exists without the route entry zebra needs to
check for it before access in the debug message

rib_process_result (ctx=0x55eafd9ff350) at zebra/zebra_rib.c:2330

Was introduced with -
        https://gitlab-master.nvidia.com/nbu-sws/CL/FRR/frr/-/merge_requests/523

Ticket:#3600655
Testing Done: Failing test passed with no crash

NOTE: this commit will not be required to be carried forward as
      if upstreamed we will include this fix in the original commit.
Signed-off-by: Ashwini Reddy <ashred@nvidia.com>

[CUMULUS] zebra: core on handling rejected routes

Changes:
   -  validate prefix before installation
   -  when the rejected list is done delete the list and explicitly set to NULL

Ticket:#3593844

Testing Done: Ran the test suite that failed and UT on
re-programming rejected routes

NOTE: this commit will not be required to be carried forward as
if upstreamed we will include this fix in the original commit.

Signed-off-by: Ashwini Reddy <ashred@nvidia.com>