pbrd: Fix for displaying DSCP and ECN only when configured
RC:
According to RFC rfc3168 section 5, the ECN field in the IP header has two bits,
allowing for four ECN codepoints from '00' to '11'. DSCP/ECN with a value of 0 should be treated
like any other value.

Issue:
Currently, however, we are displaying DSCP 0 or ECN 0 in pbr map without the user configuring it.

Fix:
We are introducing a user-set flag, which will keep track of the user-configured DSCP and ECN match.

Ticket: #3526629

Signed-off-by: Rajesh Varatharaj <rvaratharaj@nvidia.com>