tools: [Deb 12] Update to LD_PRELOAD liblttng-ust-fork.so.1 for Debian 12 LTTng support
Issue:
FRR currently utilizes LTTng version 2.10, which has been updated to version 2.13.5-1.
This update needs a change in the library, requiring an update on the FRR side to make LTTng working.

Fix:
Updated the LTTng library dependency from liblttng-ust.so.0 to liblttng-ust-fork.so.1 to align with the new version requirements.

UT:
 unit testing to confirm LTTng's operational status with the new library:
- System: Linux mlx-2410ba1-02 6.1.0-cl-1-amd64 #1 SMP PREEMPT_DYNAMIC Debian 6.1.38-4+cl5.9.0u1 (2023-08-11) x86_64 GNU/Linux
- Test Command: /usr/lib/frr/frr_babeltrace.py /var/log/lttng-traces/frr-local-zebra-onboot/
/usr/lib/frr/frr_babeltrace.py /var/log/lttng-traces/frr-local-onboot/
- Results: Confirmed successful interface updates and address additions, indicating LTTng is functioning as expected with the updated library.

Ticket: #3735825
Signed-off-by: Rajesh Varatharaj <rvaratharaj@nvidia.com>