bgpd: [SOO] soo rt and rt with soo pi bitmap apis
SOO route has 2 bitmaps
- selected path info (pi) bitmap : what is selected in bgp after best
  path of 'SOO Route'
- installed path info (pi) bitmap: what is installed in zebra as part of
  SOO NHG

Route with SOO has 1 bitmap
- selected path info (pi) bitmap: what is selected in bgp after best
  path of 'Route with SOO'

All the 'Route with SOO' are programmed with 'SOO NHG' which is a result
of processing 'SOO route'.
A SOO NHG is nothing but a NextHop group allocated and maintained in BGP
as a result of processing 'SOO route'.

To determine which ECMP paths to install to Zebra as part of 'SOO NHG'
we need the following APIs

- Check if 'SoO route' installed pi bitmap is a subset of 'route with SoO' pi bitmap
- Check if 'SoO route' selected pi bitmap is a subset of 'route with SoO' pi bitmap
- Check if SOO route path info bitmap is subset of path info bitmap of "all" the routes
  with SOO. This function walks all the "route with SOO" and checks if "SOO route" path
   info bitmap is a subset of each one of them

Ticket: #3951671

Signed-off-by: Karthikeya Venkat Muppalla <kmuppalla@nvidia.com>