staticd: [SRv6] Allocate SRv6 SIDs after connecting with SID Manager
On a heavily loaded system, it may take some time to establish a
connection between staticd and SRv6 SID Manager.
As a result, staticd might try to allocate SIDs in SID Manager before the
connection is fully established, leading to a failure.

This issue is reflected in the log with the following error message:

```
2025/03/04 15:24:01 STATIC: [HR66R-TWQYD][EC 100663302] srv6_manager_get_locator: invalid zclient socket
```

In this scenario, SIDs are still kept in the staticd configuration,
but not allocated and not installed in the RIB.

This commit fixes the issue by forcing staticd to retry to allocate
the SIDs once the connection with SID Manager is successfully
established.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>