bgpd: [GR] Reduce select-deferral timer's default value
Currently default time for GR select-deferral timer is 240secs.
Incase if some of the directly connected and multihop peers don't
comeup at all, then 240s is a lot of time to wait and this might
cause stale-path timer to time out in multihop cases.

Reducing this timer's default value to 120 seconds

Ticket:#4395761

Signed-off-by: Pooja Jagadeesh Doijode <pdoijode@nvidia.com>