zebra: Re-install all route entries upon kernel NHG action
When operations such as "ip nexthop flush" happens, kernel cleans up all
the route + nexthop entries but notifies zebra only about the nexthop
delete event.

Currently zebra code has a handling to reinstall the NHG (with a log
that there are routes still pointing to that NHG). But the problem is
that the routes are not reinstalled resulting in zebra and kernel being
out of sync.

Fix here is to maintain a rb tree of all route-entries which uses a nhe
with the head of the tree in the struct nhg_hash_entry. Once nhe is
installed (currently only when trigger is from kernel), we walk this
tree and reinstall all the re's which were SELECTED for install.
 - rib_unlink -> Deletes the route entry from NHG
 - rib_process_result
     * Deletes the route entry from NHG in case of DELETE
     * Deletes the similar/same route entry and add new entry in case of
       ADD/UPDATE
 - zebra_nhg_dplane_result
    * Walks all re's in the nhe->re_tree and re-install the ones which
      are marked SELECTED.
    * The non-SELECTED ones are deleted(if they exist)

For future purposes, we can use this nhe->re_head tree as per needed.

Signed-off-by: Rajasekar Raja <rajasekarr@nvidia.com>