zebra: [GR] z_gr_ctx.re points to freed mem during warmboot with wait-for-install
Problem:
With wait-for-install enabled, when system undergoes warmboot
z_gr_ctx.re which is used to store the last BGP route is pointing to
freed memory. This causes zebra to crash once in a while.

Fix:
Instead of accessing z_gr_ctx.re's flags/status which could be
junk (since this re could be freed already) compare the last rn's
(z_gr_ctx.rn) selected fib route to z_gr_ctx.re first. If they're the
same then check the status of re. If they are not the same, then update
the last re pointer (z_gr_ctx.re) to point to the selected_fib if
it's a installed BGP route and reinstall it as last route.

Ticket:#4411693
Signed-off-by: Pooja Jagadeesh Doijode <pdoijode@nvidia.com>