zebra: [SRv6] Provide SID value when sending SRv6 SID release notify message
When a daemon asks SID Manager to release a SID, SID Manager calls
static_zebra_srv6_sid_notify() to notify the daemon if the SID release
was successful or there was an error.

Currently, when calling static_zebra_srv6_sid_notify(), SID Manager
does not provide the SID value, which makes difficult for the daemon to
figure out which SID the notification refers to.

The problem can be clearly observed by enabling `debug static srv6`,
removing a SID in STATIC and looking at STATIC logs.

Before the change, STATIC receives a notify message from SID Manager
that does not contain a sid_value "::" :

2025/06/07 10:15:32.204664 STATIC: [J9AAH-6B3EE] static_zebra_srv6_sid_notify: received SRv6 SID notify: ctx End.DT6 vrf_id 4 sid_value :: sid_func 0 note ZAPI_SRV6_SID_RELEASED
2025/06/07 10:15:32.204691 STATIC: [HMKQ2-9A3HV] static_zebra_srv6_sid_notify: SRv6 SID :: End.DT6 vrf_id 4: RELEASED

After the change, the notify message provides an indication of which
sid_value has been released "fcbb:bbbb:1:fe20::" :

2025/06/07 10:13:35.840414 STATIC: [J9AAH-6B3EE] static_zebra_srv6_sid_notify: received SRv6 SID notify: ctx End.DT6 vrf_id 4 sid_value fcbb:bbbb:1:fe20:: sid_func 0 note ZAPI_SRV6_SID_RELEASED
2025/06/07 10:13:35.840445 STATIC: [HMKQ2-9A3HV] static_zebra_srv6_sid_notify: SRv6 SID fcbb:bbbb:1:fe20:: End.DT6 vrf_id 4: RELEASED

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>