zebra: fix check for singleton nhg
Issue:

Have NHG Z, with NHG A, B and C as members.
Now we bring down interface A, followed by interface B.
We rightly see that the NHG Z in kernel is programmed with only active member NHG.
Now when we bring back interface B, we see that NHG Z in kernel still has only NHG C as its member.
NHG B is not programmed in kernel as member of NHG Z.
Now when we bring up interface A, we see all the member NHGs are programmed in kernel for NHG Z.

Root Cause:

In zebra, on if_down, when the dependent nhgs for the interface are marked as invalid,
we are not clearing the NEXTHOP_FLAG_ACTIVE flag for nexthop in singleton NHG due to the incorrect
IF condition. The singleton NHG will be marked as invalid, but the nexthop is still marked active.

Now on if_up event(interface B), we mark the singleton NHGs associated with interface as valid,
and mark all their dependent NHGs(NHG Z) for reinstall.
Once the singleton NHG(NHG B) is installed in kernel, we will try to install all its dependent NHGs(NHG Z).
Before installing NHG Z, we iterate through all its singleton NHGs and mark them valid if
they have a nexthop with NEXTHOP_FLAG_ACTIVE flag set(Now NHG A is marked valid)
and install singleton NHGs which are not installed in kernel.

Now zebra will try to program NHG A to kernel. But will fail since interface A is not up.
Zebra will also program NHG Z with all 3 members A,B and C.
This will again fail since NHG A is not programmed in kernel.
We hence see NHG Z in kernel having only NHG C, even though interface B is up.

Fix:

Change IF condition in if_down code, to rightly mark NH in singleton NHG as inactive.

Ticket: #4465586
Signed-off-by: Aprathi K <aprathik@nvidia.com>