[CUMULUS] zebra: do not reuse nhg on local link flaps
Root cause:
Issue: 7+ seconds traffic loss observed when spine11 is rebooted.

Topology:
- leaf 11 receiving 80k route from spine11 (8 paths) and spine12 (8 paths)
- ecmp is 16-way ecmp
- Traffic is going from leaf11 to spine11 and spine12

1. In steady state, all 80k routes on leaf11 use NHG-A with 16-way ECMP.
2. When spine11 is rebooted (going down), 8 nexthops are marked as INACTIVE
   in NHG-A and depending on the sequence of events which ever gets
   processed first (link down events in zerba or bgp ADJCHANGE due to
   CEASE notification recieved from spine11), NHG-B could be created.
   We end up in a state where some routes are on NHG-A (8 ACTIVE + 8
   INACTIVE) and remaining on NHG-B (8 ACTIVE) even though NHG-A and NHG-B
   have exactly the same nexthops from kernel perspective as zerba only
   installs ACTIVE paths.
3. When spine11 is rebooted (coming up), up on recieving link up
   notificaton, the current logic in zebra is to mark nexthops in NHG-A
   from INACTIVE to ACTIVE and there by 16-way ECMP which is programmed
   in the kernel. This will lead to a scenario where traffic is
   prematurely forwarded to all the 80k routes on 16-way ECMP even
   before leaf11 has recieving all those routes from BGP of spine11.
   Since spine11 is coming up afresh, it has to install all the 80k
   routes it recieves from its upstream devices (superpsine in this
   case) or another leaf(from where route is originating). If leaf11
   switches to NHG-A with 16 paths before spine11 could program its
   hardware, we see that traffic loss. I confirmed with switch software
   table dump ADD time stamps that this is the case where NHG-A is moving
   to 16-way ECMP much earlier.

Fix:

For now, we do not re-use the NHG-A which is marked as INACTIVE when local
link goes down and there by forcing all routes to move to NHG-B in step
(2) above. In step (3) when spine11 comes up, since all routes are on
NHG-B (8 way ECMP), only when per route update comes from BGP, the
routes are moved one after another to NHG-B. So all the routes slowly
transition from NHG-B back to NHG-A.

The zebra_nhg_rib_compare_old_nhe() is the function which tries to
re-use NHG-A for routes to avoid unnecessary new NHG creation and moving
routes between NHGs, but we are reverting that behavior to fix the above
traffic loss issue.

A longer term fix is under discussion to retain routes on NHG-A and
safely transaition NHG-A from 8-way to 16-way ECMP based on some timer
logic so there is no traffic loss and unncessary creation of NHGs.

Unit test:
Tested spine reboot, partial links down/up, all links down/up - no
traffic loss observed with the fix.

Ticket: #4457165

Signed-off-by: Karthikeya Venkat Muppalla <kmuppalla@nvidia.com>