[CUMULUS] zebra: Set default NHG re-use time to 5secs
Currently when all route reference to a NHG is removed, we retain the
NHG for default 180 secs with a hope that it might be re-used. Until
this 180secs, FRR does not delete the NHG from kernel.

Setting NHG re-use time to 5secs default in Cumulus.

NOTE: Not to be carried forward. This is an internal only patch and has
to go hand in hand with MR #1197 where NHG re-use for link flap is
disabled.

Ticket :#4535938

Signed-off-by: Rajasekar Raja <rajasekarr@nvidia.com>