bgpd: incorrect mpath_count when displaying a prefix
Problem:
Multipath count for a prefix was displayed as 4 when there were only
3 paths for that prefix.

Fix:
BGP saves multipath count for a prefix in it's bestapath's mpath.
Modified the code to fetch the multipath count from bestpaths'
mapath pointer. Multipath count will be displayed as 0 when there's no
bestpath for the prefix

Ticket: #4534499